more donor related fixes / error handling, cleanup, refactoring
