+++ rvas-core_v0.3/pom.xml
@@ -75,8 +75,9 @@
     </dependencies>
 
     <build>
-        <finalName>rvas-core-0.3.5-325</finalName>
+        <finalName>rvas-core-0.3.5-326</finalName>
         <plugins>
+
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-eclipse-plugin</artifactId>
