+++ rvas-core_v0.3/src/main/java/core/backend/cmd/DupeHand.java
@@ -1,8 +1,9 @@
-package core.backend;
+package core.backend.cmd;
 
+import core.backend.Config;
 import core.data.PlayerMeta;
 import core.frontend.ChatPrint;
-import core.backend.anno.Critical;
+import core.backend.ex.Critical;
 
 import java.util.*;
 import java.util.stream.IntStream;
