+++ rvas-core_v0.3/src/main/java/core/commands/Kill.java
@@ -2,7 +2,7 @@ package core.commands;
 
 import core.tasks.Analytics;
 import core.data.PlayerMeta;
-import core.backend.anno.Critical;
+import core.backend.ex.Critical;
 
 import java.util.Objects;
 import org.bukkit.Bukkit;
