+++ rvas-core_v0.3/src/main/java/core/commands/Local.java
@@ -2,7 +2,7 @@ package core.commands;
 
 import core.data.PlayerMeta;
 import core.frontend.ChatPrint;
-import core.commands.restricted.Admin;
+import core.commands.op.Admin;
 
 import java.util.ArrayList;
 import java.util.Arrays;
