+++ rvas-core_v0.3/src/main/java/core/commands/Redeem.java
@@ -1,11 +1,32 @@
 package core.commands;
 
+/* *
+ *  About: A command for non-ops to register themselves as
+ * 		Donor objects when they have a valid donation key
+ *
+ *  LICENSE: AGPLv3 (https://www.gnu.org/licenses/agpl-3.0.en.html)
+ *  Copyright (C) 2021  Lysergik Productions (https://github.com/LysergikProductions)
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU Affero General Public License as
+ *  published by the Free Software Foundation, either version 3 of the
+ *  License, or (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU Affero General Public License for more details.
+ *
+ *  You should have received a copy of the GNU Affero General Public License
+ *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ *
+ * */
+
 import core.frontend.ChatPrint;
 import core.data.DonationManager;
-import core.backend.anno.Critical;
+import core.backend.ex.Critical;
 
 import java.io.IOException;
-
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 import net.md_5.bungee.api.chat.TextComponent;
