+++ rvas-core_v0.3/src/main/java/core/commands/Reply.java
@@ -3,8 +3,8 @@ package core.commands;
 import core.data.PlayerMeta;
 import core.tasks.Analytics;
 import core.frontend.ChatPrint;
-import core.commands.restricted.Admin;
-import core.backend.anno.Critical;
+import core.commands.op.Admin;
+import core.backend.ex.Critical;
 
 import java.util.UUID;
 import java.util.Arrays;
