+++ rvas-core_v0.3/src/main/java/core/commands/Tps.java
@@ -4,7 +4,7 @@ import core.data.PlayerMeta;
 import core.tasks.Analytics;
 import core.frontend.ChatPrint;
 import core.backend.LagProcessor;
-import core.backend.anno.Critical;
+import core.backend.ex.Critical;
 
 import java.text.DecimalFormat;
 import org.apache.commons.lang.math.IntRange;
