+++ rvas-core_v0.3/src/main/java/core/commands/op/Global.java
@@ -1,4 +1,4 @@
-package core.commands.restricted;
+package core.commands.op;
 
 /* *
  *
