+++ rvas-core_v0.3/src/main/java/core/commands/op/Info.java
@@ -1,8 +1,8 @@
-package core.commands.restricted;
+package core.commands.op;
 
 /* *
- *  About: A command for ops to toggle the configured per-user chat cool-down
- *
+ *  About: A command for ops to see current session data
+ * 
  *  LICENSE: AGPLv3 (https://www.gnu.org/licenses/agpl-3.0.en.html)
  *  Copyright (C) 2021  Lysergik Productions (https://github.com/LysergikProductions)
  *
@@ -18,33 +18,47 @@ package core.commands.restricted;
  *
  *  You should have received a copy of the GNU Affero General Public License
  *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
- *
+ * 
  * */
 
-import core.events.ChatListener;
+import core.frontend.ChatPrint;
+import core.backend.ServerMeta;
+import core.backend.utils.Util;
+
+import core.events.ChunkManager;
+import core.events.BlockListener;
+import core.events.SpawnController;
+
 import org.bukkit.entity.Player;
-import core.backend.anno.Critical;
+import net.md_5.bungee.api.chat.TextComponent;
 
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
 import org.jetbrains.annotations.NotNull;
 
-@Critical
-public class SlowChat implements CommandExecutor {
-	static String msg;
+public class Info implements CommandExecutor {
 	
 	@Override
 	public boolean onCommand(@NotNull CommandSender sender, @NotNull Command command, @NotNull String s, String[] args) {
 		
 		Player player = (Player) sender;
 		if (!player.isOp()) return false;
+			
+		String humanUptime = Util.durationFormat(ServerMeta.getUptime());
 		
-		if (args.length == 0) ChatListener.slowChatEnabled = !ChatListener.slowChatEnabled;
+		TextComponent head = new TextComponent("--- Session Stats ---");
+		head.setColor(ChatPrint.primary); head.setBold(true);
 		
-		if (ChatListener.slowChatEnabled) msg = "enabled!"; else msg = "disabled!";
+		player.sendMessage(head.toLegacyText());
+		player.sendMessage("Uptime: " + humanUptime);
+		player.sendMessage("New Chunks: " + ChunkManager.newCount);
+		player.sendMessage("New Players: " + SpawnController.sessionNewPlayers);
+		player.sendMessage("Total Respawns: " + SpawnController.sessionTotalRespawns);
 		
-		player.sendMessage("Slow chat is " + msg);
+		player.sendMessage("Bedrock Placed: " + BlockListener.placedBedrockCounter);
+		player.sendMessage("Bedrock Broken: " + BlockListener.brokenBedrockCounter);
+		// TODO : add total BlockPlaceEvent's and BlockBreakEvent's
 		
 		return true;
 	}
