+++ rvas-core_v0.3/src/main/java/core/commands/op/Prison.java
@@ -1,10 +1,10 @@
-package core.commands.restricted;
+package core.commands.op;
 
 import core.backend.Config;
 import core.frontend.ChatPrint;
 import core.data.PlayerMeta;
 import core.data.PrisonerManager;
-import core.backend.anno.Critical;
+import core.backend.ex.Critical;
 
 import java.util.*;
 import org.bukkit.Bukkit;
