+++ rvas-core_v0.3/src/main/java/core/commands/op/Speeds.java
@@ -1,11 +1,11 @@
-package core.commands.restricted;
+package core.commands.op;
 
 import core.Main;
 import core.frontend.ChatPrint;
 import core.backend.utils.Util;
 import core.data.objects.Pair;
 import core.events.SpeedLimiter;
-import core.backend.anno.Critical;
+import core.backend.ex.Critical;
 
 import java.util.*;
 import java.util.logging.Level;
