+++ rvas-core_v0.3/src/main/java/core/data/StatsManager.java
@@ -1,7 +1,6 @@
 package core.data;
 
 /* *
- * 
  *  About: Reads, writes, and mutates StatsContainer objects
  * 
  *  LICENSE: AGPLv3 (https://www.gnu.org/licenses/agpl-3.0.en.html)
