+++ rvas-core_v0.3/src/main/java/core/events/ChatListener.java
@@ -4,8 +4,8 @@ import core.backend.Config;
 import core.frontend.ChatPrint;
 import core.commands.AFK;
 import core.commands.Message;
-import core.commands.restricted.Admin;
-import core.backend.anno.Critical;
+import core.commands.op.Admin;
+import core.backend.ex.Critical;
 
 import core.data.PlayerMeta;
 import core.data.objects.Donor;
