+++ rvas-core_v0.3/src/main/java/core/tasks/AutoAnnouncer.java
@@ -1,7 +1,6 @@
 package core.tasks;
 
 /* *
- *
  *  About: Announce string randomly selected from
  * 			announcements.txt as a synced scheduled task
  *
@@ -26,8 +25,8 @@ package core.tasks;
 import core.backend.Config;
 import core.backend.Scheduler;
 import core.frontend.ChatPrint;
-import core.commands.restricted.Check;
-import core.backend.anno.Critical;
+import core.commands.op.Check;
+import core.backend.ex.Critical;
 
 import java.util.*;
 import java.nio.file.Files;
@@ -42,7 +41,6 @@ import net.md_5.bungee.api.chat.TextComponent;
 import net.md_5.bungee.api.chat.hover.content.Text;
 
 @Critical
-@SuppressWarnings("deprecation")
 public class AutoAnnouncer extends TimerTask {
 
 	String lastAnnouncement = "";
@@ -69,6 +67,7 @@ public class AutoAnnouncer extends TimerTask {
 	}
 
 	@Override
+	@SuppressWarnings("deprecation")
 	public void run() {
 		if (Config.getValue("announcer.enabled").equals("false")) return;
