+++ rvas-core_v0.3/src/main/java/core/tasks/OnTick.java
@@ -2,7 +2,7 @@ package core.tasks;
 
 import core.backend.Scheduler;
 import core.commands.VoteMute;
-import core.backend.anno.Critical;
+import core.backend.ex.Critical;
 
 import java.util.TimerTask;
